third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-dom-width-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' attribute of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.getAttribute('width') is "100"
PASS filterElement.getAttribute('width') is "200"
PASS successfullyParsed is true

TEST COMPLETE
